+++ legadilo/users/views/__init__.py
@@ -6,6 +6,8 @@ from .manage_tokens_views import delete_token_view, manage_tokens_view
 from .notifications_views import list_notifications_view
 from .user_views import (
     delete_account_view,
+    logout_view,
+    signup_view,
     user_redirect_view,
     user_update_settings_view,
     user_update_view,
@@ -15,7 +17,9 @@ __all__ = [
     "delete_account_view",
     "delete_token_view",
     "list_notifications_view",
+    "logout_view",
     "manage_tokens_view",
+    "signup_view",
     "user_redirect_view",
     "user_update_settings_view",
     "user_update_view",