+++ CHANGELOG.md
@@ -19,6 +19,7 @@ SPDX-License-Identifier: CC-BY-SA-4.0
   - Sessions are now cleaned as part of the `clean_users` command. `clearsessions` cannot be used anymore.
 - Enable account deletion in the account parameters.
 - Inactive accounts are deleted after several notification emails. Both the deletion and the notification are done by the `clean_users` command.
+- Allow user to specify their language. Use the browser one by default and English as fallback.
 
 ## 25.12.5